backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-multipart>=0.0.9
google-generativeai>=0.7.0
google-cloud-speech>=2.26.0
Pillow>=10.3.0
packaging>=24.0

